+++ src/main/java/com/uber/rss/common/MapTaskRssInfo.java
@@ -23,18 +23,18 @@ import java.util.Base64;
  * This class store track information for a map task, e.g. task attempt id.
  */
 public class MapTaskRssInfo {
-  private final int mapId;
+  private final long mapId;
   private final long taskAttemptId;
   private final int numRssServers;
 
   public void serialize(ByteBuf buf) {
-    buf.writeInt(mapId);
+    buf.writeLong(mapId);
     buf.writeLong(taskAttemptId);
     buf.writeInt(numRssServers);
   }
 
   public static MapTaskRssInfo deserialize(ByteBuf buf) {
-    int mapId = buf.readInt();
+    long mapId = buf.readLong();
     long taskAttemptId = buf.readLong();
     int numRssServers = buf.readInt();
     return new MapTaskRssInfo(mapId, taskAttemptId, numRssServers);
@@ -66,13 +66,13 @@ public class MapTaskRssInfo {
     }
   }
 
-  public MapTaskRssInfo(int mapId, long taskAttemptId, int numRssServers) {
+  public MapTaskRssInfo(long mapId, long taskAttemptId, int numRssServers) {
     this.mapId = mapId;
     this.taskAttemptId = taskAttemptId;
     this.numRssServers = numRssServers;
   }
 
-  public int getMapId() {
+  public long getMapId() {
     return mapId;
   }
